Refactor: break up `user.go` to separate files
