+++ enterprise/internal/scim/user.go
@@ -3,22 +3,13 @@ package scim
 import (
 	"context"
 	"net/http"
-	"strconv"
-	"strings"
-	"time"
 
 	"github.com/elimity-com/scim"
-	scimerrors "github.com/elimity-com/scim/errors"
 	"github.com/elimity-com/scim/optional"
 	"github.com/elimity-com/scim/schema"
 	"github.com/sourcegraph/log"
-	"github.com/sourcegraph/sourcegraph/enterprise/internal/scim/filter"
-	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
-	"github.com/sourcegraph/sourcegraph/internal/types"
-	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
 // UserResourceHandler implements the scim.ResourceHandler interface for users.
@@ -41,286 +32,6 @@ func NewUserResourceHandler(ctx context.Context, observationCtx *observation.Con
 	}
 }
 
-// Create stores given attributes. Returns a resource with the attributes that are stored and a (new) unique identifier.
-func (h *UserResourceHandler) Create(r *http.Request, attributes scim.ResourceAttributes) (scim.Resource, error) {
-	// Get external ID, primary email, username, and display name
-	optionalExternalID := getOptionalExternalID(attributes)
-	primaryEmail := extractPrimaryEmail(attributes)
-	if primaryEmail == "" {
-		return scim.Resource{}, scimerrors.ScimErrorBadParams([]string{"emails missing"})
-	}
-	username := extractUsername(attributes)
-	displayName := extractDisplayName(attributes)
-
-	// Create user (with or without external ID)
-	// TODO: Use NewSCIMUser instead of NewUser?
-	newUser := database.NewUser{
-		Email:           primaryEmail,
-		Username:        username,
-		DisplayName:     displayName,
-		EmailIsVerified: true,
-	}
-	var user *types.User
-	var err error
-	if optionalExternalID.Present() {
-		accountSpec := extsvc.AccountSpec{
-			ServiceType: "scim",
-			// TODO: provide proper service ID
-			ServiceID: "TODO",
-			AccountID: optionalExternalID.Value(),
-		}
-		user, err = h.db.UserExternalAccounts().CreateUserAndSave(r.Context(), newUser, accountSpec, extsvc.AccountData{})
-	} else {
-		user, err = h.db.Users().Create(r.Context(), newUser)
-	}
-	if err != nil {
-		if dbErr, ok := containsDBError(err); ok {
-			if code := dbErr.Code(); code == database.ErrorCodeUsernameExists || code == database.ErrorCodeEmailExists {
-				return scim.Resource{}, scimerrors.ScimError{Status: http.StatusConflict, Detail: err.Error()}
-			}
-		}
-		return scim.Resource{}, scimerrors.ScimError{Status: http.StatusInternalServerError, Detail: err.Error()}
-	}
-
-	var now = time.Now()
-
-	return scim.Resource{
-		ID:         strconv.Itoa(int(user.ID)),
-		ExternalID: optionalExternalID,
-		Attributes: attributes,
-		Meta: scim.Meta{
-			Created:      &now,
-			LastModified: &now,
-		},
-	}, nil
-}
-
-// extractPrimaryEmail extracts the primary email address from the given attributes.
-// Tries to get the (first) email address marked as primary, otherwise uses the first email address it finds.
-func extractPrimaryEmail(attributes scim.ResourceAttributes) (primaryEmail string) {
-	if attributes["emails"] == nil {
-		return
-	}
-	emails := attributes["emails"].([]interface{})
-	for _, emailRaw := range emails {
-		email := emailRaw.(map[string]interface{})
-		if email["primary"] == true {
-			primaryEmail = email["value"].(string)
-			break
-		}
-	}
-	if primaryEmail == "" && len(emails) > 0 {
-		primaryEmail = emails[0].(map[string]interface{})["value"].(string)
-	}
-	return
-}
-
-// extractUsername extracts the username from the given attributes.
-func extractUsername(attributes scim.ResourceAttributes) (username string) {
-	if attributes["userName"] != nil {
-		username = attributes["userName"].(string)
-	}
-	return
-}
-
-// extractDisplayName extracts the user's display name from the given attributes.
-func extractDisplayName(attributes scim.ResourceAttributes) (displayName string) {
-	if attributes["displayName"] != nil {
-		displayName = attributes["displayName"].(string)
-	} else if attributes["name"] != nil {
-		name := attributes["name"].(map[string]interface{})
-		if name["formatted"] != nil {
-			displayName = name["formatted"].(string)
-		} else if name["givenName"] != nil && name["familyName"] != nil {
-			displayName = name["givenName"].(string) + " " + name["familyName"].(string)
-		}
-	} else if attributes["userName"] != nil {
-		displayName = attributes["userName"].(string)
-	}
-	return
-}
-
-// containsDBError returns true if the given error contains at least one database.ErrCannotCreateUser.
-// It also returns the first such error.
-func containsDBError(err error) (database.ErrCannotCreateUser, bool) {
-	if err == nil {
-		return database.ErrCannotCreateUser{}, false
-	}
-	if _, ok := err.(database.ErrCannotCreateUser); ok {
-		return err.(database.ErrCannotCreateUser), true
-	}
-
-	// Handle multiError
-	if multiErr, ok := err.(errors.MultiError); ok {
-		for _, err := range multiErr.Errors() {
-			if _, ok := err.(database.ErrCannotCreateUser); ok {
-				return err.(database.ErrCannotCreateUser), true
-			}
-		}
-	}
-
-	return database.ErrCannotCreateUser{}, false
-}
-
-// getOptionalExternalID extracts the external identifier of the given attributes.
-func getOptionalExternalID(attributes scim.ResourceAttributes) optional.String {
-	if eID, ok := attributes["externalId"]; ok {
-		if externalID, ok := eID.(string); ok {
-			return optional.NewString(externalID)
-		}
-	}
-	return optional.String{}
-}
-
-// Get returns the resource corresponding with the given identifier.
-func (h *UserResourceHandler) Get(r *http.Request, idStr string) (scim.Resource, error) {
-	id, err := strconv.ParseInt(idStr, 10, 32)
-	if err != nil {
-		return scim.Resource{}, errors.New("invalid id")
-	}
-
-	// Get users
-	users, err := h.db.Users().ListForSCIM(r.Context(), &database.UsersListOptions{
-		UserIDs: []int32{int32(id)},
-	})
-	if err != nil {
-		return scim.Resource{}, scimerrors.ScimError{Status: http.StatusInternalServerError, Detail: err.Error()}
-	}
-	if len(users) == 0 {
-		return scim.Resource{}, scimerrors.ScimErrorResourceNotFound(idStr)
-	}
-
-	resource := h.convertUserToSCIMResource(users[0])
-
-	return resource, nil
-}
-
-// GetAll returns a paginated list of resources.
-// An empty list of resources will be represented as `null` in the JSON response if `nil` is assigned to the
-// Page.Resources. Otherwise, if an empty slice is assigned, an empty list will be represented as `[]`.
-func (h *UserResourceHandler) GetAll(r *http.Request, params scim.ListRequestParams) (scim.Page, error) {
-	var totalCount int
-	var resources []scim.Resource
-	var err error
-
-	if params.Filter == nil {
-		totalCount, resources, err = h.getAllFromDB(r, params.StartIndex, &params.Count)
-	} else {
-		extensionSchemas := make([]schema.Schema, 0, len(h.schemaExtensions))
-		for _, ext := range h.schemaExtensions {
-			extensionSchemas = append(extensionSchemas, ext.Schema)
-		}
-		validator := filter.NewFilterValidator(params.Filter, h.coreSchema, extensionSchemas...)
-
-		// Fetch all resources from the DB and then filter them here.
-		// This doesn't feel efficient, but it wasn't reasonable to implement this in SQL in the time available.
-		var allResources []scim.Resource
-		_, allResources, err = h.getAllFromDB(r, 0, nil)
-
-		for _, resource := range allResources {
-			if err := validator.PassesFilter(resource.Attributes); err != nil {
-				continue
-			}
-
-			totalCount++
-			if totalCount >= params.StartIndex && len(resources) < params.Count {
-				resources = append(resources, resource)
-			}
-		}
-	}
-	if err != nil {
-		return scim.Page{}, scimerrors.ScimError{Status: http.StatusInternalServerError, Detail: err.Error()}
-	}
-
-	return scim.Page{
-		TotalResults: totalCount,
-		Resources:    resources,
-	}, nil
-}
-
-func (h *UserResourceHandler) getAllFromDB(r *http.Request, startIndex int, count *int) (totalCount int, resources []scim.Resource, err error) {
-	// Calculate offset
-	var offset int
-	if startIndex > 0 {
-		offset = startIndex - 1
-	}
-
-	// Get users and convert them to SCIM resources
-	var opt = &database.UsersListOptions{}
-	if count != nil {
-		opt = &database.UsersListOptions{
-			LimitOffset: &database.LimitOffset{Limit: *count, Offset: offset},
-		}
-	}
-	users, err := h.db.Users().ListForSCIM(r.Context(), opt)
-	if err != nil {
-		return
-	}
-	resources = make([]scim.Resource, 0, len(users))
-	for _, user := range users {
-		resources = append(resources, h.convertUserToSCIMResource(user))
-	}
-
-	// Get total count
-	if count == nil {
-		totalCount = len(users)
-	} else {
-		totalCount, err = h.db.Users().Count(r.Context(), &database.UsersListOptions{})
-	}
-
-	return
-}
-
-// convertUserToSCIMResource converts a Sourcegraph user to a SCIM resource.
-func (h *UserResourceHandler) convertUserToSCIMResource(user *types.UserForSCIM) scim.Resource {
-	// Convert names
-	firstName, middleName, lastName := displayNameToPieces(user.DisplayName)
-
-	// Convert external ID
-	externalIDOptional := optional.String{}
-	if user.SCIMExternalID != "" {
-		externalIDOptional = optional.NewString(user.SCIMExternalID)
-	}
-
-	// Convert emails
-	emailMap := make([]interface{}, 0, len(user.Emails))
-	for _, email := range user.Emails {
-		emailMap = append(emailMap, map[string]interface{}{"value": email})
-	}
-
-	return scim.Resource{
-		ID:         strconv.FormatInt(int64(user.ID), 10),
-		ExternalID: externalIDOptional,
-		Attributes: scim.ResourceAttributes{
-			"userName":   user.Username,
-			"externalId": user.SCIMExternalID,
-			"name": map[string]interface{}{
-				"givenName":  firstName,
-				"middleName": middleName,
-				"familyName": lastName,
-				"formatted":  user.DisplayName,
-			},
-			"displayName": user.DisplayName,
-			"emails":      emailMap,
-			"active":      true,
-		},
-	}
-}
-
-// displayNameToPieces splits a display name into first, middle, and last name.
-func displayNameToPieces(displayName string) (first, middle, last string) {
-	pieces := strings.Fields(displayName)
-	switch len(pieces) {
-	case 0:
-		return "", "", ""
-	case 1:
-		return pieces[0], "", ""
-	case 2:
-		return pieces[0], "", pieces[1]
-	default:
-		return pieces[0], strings.Join(pieces[1:len(pieces)-1], " "), pieces[len(pieces)-1]
-	}
-}
 
 // Replace replaces ALL existing attributes of the resource with given identifier. Given attributes that are empty
 // are to be deleted. Returns a resource with the attributes that are stored.
@@ -332,101 +43,6 @@ func (h *UserResourceHandler) Replace(r *http.Request, id string, attributes sci
 	return scim.Resource{
 		ID: "123",
 	}, nil
-}
-
-// Delete removes the resource with corresponding ID.
-func (h *UserResourceHandler) Delete(r *http.Request, id string) error {
-	logger := h.observationCtx.Logger.Scoped("DeleteUsers", "SCIM delete user").With(log.String("user", id))
-
-	// Find user
-	idInt, err := strconv.Atoi(id)
-	if err != nil {
-		return errors.Wrap(err, "parse user ID")
-	}
-	users, err := h.db.Users().ListForSCIM(r.Context(), &database.UsersListOptions{
-		UserIDs: []int32{int32(idInt)},
-	})
-	if err != nil {
-		return errors.Wrap(err, "list users by IDs")
-	}
-	if len(users) == 0 {
-		logger.Info("requested users to delete do not exist")
-		return nil
-	} else if len(users) > 1 {
-		logger.Error("requested users to delete have duplicate IDs—that should not happen")
-		return errors.New("requested users to delete have duplicate IDs")
-	} else {
-		logger.Debug("attempting to delete requested users")
-	}
-	if users[0].SCIMExternalID != "" {
-		return errors.New("cannot delete user because it has no SCIM external ID set")
-	}
-
-	// Collect username, verified email addresses, and external accounts to be used for revoking user permissions.
-	revokeUserPermissionsArgsList, err := getRevokeUserPermissionArgs(r.Context(), users, h.db)
-	if err != nil {
-		return err
-	}
-
-	// Delete user
-	if err := h.db.Users().HardDelete(r.Context(), int32(idInt)); err != nil {
-		return err
-	}
-
-	// NOTE: Practically, we don't reuse the ID for any new users, and the situation of left-over pending permissions
-	// is possible but highly unlikely. Therefore, there is no need to roll back user deletion even if this step failed.
-	// This call is purely for the purpose of cleanup.
-	if err := h.db.Authz().RevokeUserPermissionsList(r.Context(), revokeUserPermissionsArgsList); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// getRevokeUserPermissionArgs returns a list of arguments for revoking user permissions.
-func getRevokeUserPermissionArgs(ctx context.Context, users []*types.UserForSCIM, db database.DB) ([]*database.RevokeUserPermissionsArgs, error) {
-	accountsList := make([][]*extsvc.Accounts, len(users))
-	var revokeUserPermissionsArgsList []*database.RevokeUserPermissionsArgs
-	for index, user := range users {
-		var accounts []*extsvc.Accounts
-
-		extAccounts, err := db.UserExternalAccounts().List(ctx, database.ExternalAccountsListOptions{UserID: user.ID})
-		if err != nil {
-			return nil, errors.Wrap(err, "list external accounts")
-		}
-		for _, acct := range extAccounts {
-			accounts = append(accounts, &extsvc.Accounts{
-				ServiceType: acct.ServiceType,
-				ServiceID:   acct.ServiceID,
-				AccountIDs:  []string{acct.AccountID},
-			})
-		}
-
-		verifiedEmails, err := db.UserEmails().ListByUser(ctx, database.UserEmailsListOptions{
-			UserID:       user.ID,
-			OnlyVerified: true,
-		})
-		if err != nil {
-			return nil, errors.Wrap(err, "list verified emails")
-		}
-		emailStrings := make([]string, len(verifiedEmails))
-		for i := range verifiedEmails {
-			emailStrings[i] = verifiedEmails[i].Email
-		}
-		accounts = append(accounts, &extsvc.Accounts{
-			ServiceType: authz.SourcegraphServiceType,
-			ServiceID:   authz.SourcegraphServiceID,
-			AccountIDs:  append(emailStrings, user.Username),
-		})
-
-		accountsList[index] = accounts
-
-		revokeUserPermissionsArgsList = append(revokeUserPermissionsArgsList, &database.RevokeUserPermissionsArgs{
-			UserID:   user.ID,
-			Accounts: accounts,
-		})
-	}
-	return revokeUserPermissionsArgsList, nil
 }
 
 // Patch update one or more attributes of a SCIM resource using a sequence of
